wacli_api/schemas.py
"""Pydantic request/response envelope models."""

from __future__ import annotations

from typing import Generic, TypeVar

from pydantic import BaseModel

T = TypeVar("T")


class ApiResponse(BaseModel, Generic[T]):
    success: bool
    data: T | None = None
    error: str | None = None


class SendTextRequest(BaseModel):
    to: str
    body: str
