pyproject.toml
[project]
name = "wacli-api"
version = "0.1.0"
description = "HTTP API wrapping the wacli WhatsApp CLI"
authors = [{name = "loup"}]
requires-python = ">=3.13,<4"
dependencies = [
    "fastapi (>=0.115)",
    "uvicorn[standard] (>=0.34)",
    "pydantic-settings (>=2.0)",
    "python-multipart (>=0.0.18)",
    "redis (>=5.0)",
    "httpx (>=0.28)",
    "fastmcp (>=2.1)",
    "sqlmodel (>=0.0.21)",
]

[project.scripts]
wacli-api = "wacli_api.main:run"

[tool.poetry]
packages = [{include = "wacli_api"}]

[tool.poetry.group.dev.dependencies]
ruff = ">=0.11"
pyright = ">=1.1"
pytest = ">=8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]

[tool.ruff.lint.per-file-ignores]
"wacli_api/routes/**" = ["B008"]  # Depends() in function defaults is standard FastAPI
"wacli_api/deps.py" = ["B008"]    # same
# UP046: PEP 695 generic syntax is incompatible with `from __future__ import annotations`
# when used with Pydantic BaseModel — the Generic[T] form is required here.
"wacli_api/schemas.py" = ["UP046"]

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"

[tool.pytest.ini_options]
testpaths = ["tests"]
